guardado de js

+++ Curso-JavaScript/pruebas-de-logica/Ejercicio2.js
@@ -4,7 +4,7 @@ Si es menor, no hay descuento.
 Imprime el total a pagar final (restando el descuento si aplica).*/
 let montoCompra = 0.0;
 const descuenro = 0.8;
- montoCompra = parseInt(prompt('ingrese el monto de la compra: '));
+ montoCompra = parseFloat(prompt('ingrese el monto de la compra: '));
  if (montoCompra > 100.00){
     let preciofinal = montoCompra * descuenro; 
     document.writeln(`<strong>el precio final es :</strong> ${preciofinal}`);
